// src/app/shared/models/payment.ts
export class Payment {
    payment_id: number;
    user_id: number;
    sum: number;
    status_id: number;
    create_time: number;
    title: string;
    pay_type: string;
    promotion: number;
    total: number;
    seen_flag: string;
    ship: string;
    address: string;
    voucher: string;
    shipfee: number;
    shipcode: string;
    note: string;
    phone: string;
    name: string;
    tinhthanh: string;
    quanhuyen: string;
    hinhthuc: string;
}